Created Strassen.py
This will Multiply two matrices of the order 2^n with both trivial and Strassen methods and compare the time taken by both of the Methods. It is Observed that the Strassen is taking more time than the trivial matrix Multiplication this is due to the fact that the overall execution time for Strassen() is more than the trivial method, but asymptotical the Strassen method is of the order O(nlg7) whereas the trivial method is of the Order O(n^3).